+++ tests/utils.py
@@ -62,6 +62,8 @@ TASK_NAME = "export_posts"
 TASK_DESCRIPTION = "Exporting posts..."
 MISC_PROGRESS_INT = 37
 APP_MODELS_JOB_FETCH = "app.models.Job.fetch"
+ADMIN_ROUTE = "/admin"
+ADMIN_USER_ROUTE = "/admin/user"
 
 
 class UserTestObject: